actions: Split out zerver.actions.alert_words.
(cherry picked from commit 241463e21570d168d98cd078b1fe9c269a23bf4b)
