+++ zerver/tests/test_soft_deactivation.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 from django.utils.timezone import now as timezone_now
 
-from zerver.lib.actions import do_add_alert_words
+from zerver.actions.alert_words import do_add_alert_words
 from zerver.lib.soft_deactivation import (
     add_missing_messages,
     do_auto_soft_deactivate_users,
